Rename package function2 to function
