+++ pkg/refactor/function/function.go
@@ -1,4 +1,4 @@
-package function2
+package function
 
 import (
 	"fmt"
