+++ pkg/refactor/function/helper.go
@@ -1,5 +1,5 @@
 // nolint
-package function2
+package function
 
 import (
 	"github.com/mattn/go-shellwords"
