+++ pkg/refactor/function/reconcile_test.go
@@ -1,4 +1,4 @@
-package function2
+package function
 
 import (
 	"fmt"
